+++ packages/datastore/__tests__/helpers.ts
@@ -4,6 +4,7 @@ import {
 	ModelInit,
 	Schema,
 	InternalSchema,
+	SchemaModel,
 	isModelAttributePrimaryKey,
 	__modelMeta__,
 } from '../src/types';
@@ -114,7 +115,7 @@ export function errorsFrom<T extends Object>(
 
 /**
  * Checks to see if a given object contains any extra, unexpected properties.
- * If any are present, it returns the list of unexpectd fields.
+ * If any are present, it returns the list of unexpected fields.
  *
  * @param data the object that MIGHT contain extra fields.
  * @param template the authorative template object.
@@ -524,6 +525,7 @@ class FakeGraphQLService {
 	public log: (channel: string, ...etc: any) => void = s => undefined;
 	public requests = [] as any[];
 	public tables = new Map<string, Map<string, any[]>>();
+	public tableDefinitions = new Map<string, SchemaModel>();
 	public PKFields = new Map<string, string[]>();
 	public observers = new Map<
 		string,
@@ -533,6 +535,7 @@ class FakeGraphQLService {
 	constructor(public schema: Schema) {
 		for (const model of Object.values(schema.models)) {
 			this.tables.set(model.name, new Map<string, any[]>());
+			this.tableDefinitions.set(model.name, model);
 			let CPKFound = false;
 			for (const attribute of model.attributes || []) {
 				if (isModelAttributePrimaryKey(attribute)) {
@@ -685,6 +688,39 @@ class FakeGraphQLService {
 		};
 	}
 
+	private makeExtraFieldInputError(tableName, operation, fields) {
+		const properOperationName = `${operation[0].toUpperCase()}${operation.substring(
+			1
+		)}`;
+		const inputName = `${properOperationName}${tableName}Input`;
+		return {
+			data: null,
+			errors: fields.map(field => ({
+				path: null,
+				locations: null,
+				message: `The variables input contains a field name '${field}' that is not defined for input object type '${inputName}'`,
+			})),
+		};
+	}
+
+	private makeOwnerFieldNullInputError(tableName, operation) {
+		// Response from AppSync console on non-existent model.
+		return {
+			path: [operation],
+			data: null,
+			errorType: 'Unauthorized',
+			errorInfo: null,
+			locations: [
+				{
+					line: 12,
+					column: 3,
+					sourceName: null,
+				},
+			],
+			message: `Not Authorized to access ${operation} on type ${tableName}`,
+		};
+	}
+
 	private disconnectedError() {
 		return {
 			data: {},
@@ -696,9 +732,83 @@ class FakeGraphQLService {
 		};
 	}
 
+	private ownerFields(tableName) {
+		const def = this.tableDefinitions.get(tableName)!;
+		const auth = def.attributes?.find(a => a.type === 'auth');
+		const ownerFields = auth?.properties?.rules
+			.map(rule => rule.ownerField)
+			.filter(f => f);
+
+		return ownerFields || ['owner'];
+	}
+
+	private writeableFields(tableName) {
+		const def = this.tableDefinitions.get(tableName)!;
+		return Object.keys(def.fields).filter(
+			field => !def.fields[field]?.isReadOnly
+		);
+	}
+
+	private identifyExtraValues(expected, actual) {
+		const extraValues: string[] = [];
+		for (const v of actual) {
+			if (!expected.includes(v)) {
+				extraValues.push(v);
+			}
+		}
+
+		return extraValues;
+	}
+
+	private validate(tableName, operation, record) {
+		// very simple validation for an observed *near*-regression from a PR right now.
+		// https://github.com/aws-amplify/amplify-js/pull/10915
+		const writeableFields = this.writeableFields(tableName);
+
+		let error: any;
+
+		switch (operation) {
+			case 'create':
+			case 'update':
+				const unexpectedFields = this.identifyExtraValues(
+					[...writeableFields, '_version'],
+					Object.keys(record)
+				);
+				if (unexpectedFields.length > 0) {
+					error = this.makeExtraFieldInputError(
+						tableName,
+						operation,
+						unexpectedFields
+					);
+				}
+				for (const ownerField of this.ownerFields(tableName)) {
+					if (record[ownerField] === null) {
+						error = this.makeOwnerFieldNullInputError(tableName, operation);
+						break;
+					}
+				}
+				break;
+			case 'delete':
+				break;
+			default:
+				// this is not a GraphQL error. it likely indicates our fake graphql
+				// service is broken.
+				throw new Error('Invalid operation. Should be unreachable.');
+		}
+
+		this.log('validate', {
+			tableName,
+			operation,
+			record,
+			error,
+		});
+
+		return error;
+	}
+
 	private populatedFields(record) {
 		return Object.fromEntries(
-			Object.entries(record).filter(([key, value]) => value)
+			Object.entries(record).filter(([key, value]) => value !== undefined)
 		);
 	}
 
@@ -757,7 +867,15 @@ class FakeGraphQLService {
 
 		this.log('Parsed Request', parsed);
 
-		this.requests.push({ query, variables, authMode, authToken });
+		this.requests.push({
+			query,
+			variables,
+			authMode,
+			authToken,
+			operation,
+			tableName,
+			type,
+		});
 		let data;
 		let errors = [] as any;
 
@@ -782,7 +900,13 @@ class FakeGraphQLService {
 			const record = variables.input;
 			if (type === 'create') {
 				const existing = table.get(this.getPK(tableName, record));
-				if (existing) {
+				const validationError = this.validate(tableName, 'create', record);
+				if (validationError) {
+					data = {
+						[selection]: null,
+					};
+					errors = [validationError];
+				} else if (existing) {
 					data = {
 						[selection]: null,
 					};
@@ -802,7 +926,13 @@ class FakeGraphQLService {
 				// Simulate update using the default (AUTO_MERGE) for now.
 				// NOTE: We're not doing list/set merging. :o
 				const existing = table.get(this.getPK(tableName, record));
-				if (!existing) {
+				const validationError = this.validate(tableName, 'update', record);
+				if (validationError) {
+					data = {
+						[selection]: null,
+					};
+					errors = [validationError];
+				} else if (!existing) {
 					data = {
 						[selection]: null,
 					};
@@ -816,8 +946,15 @@ class FakeGraphQLService {
 				}
 			} else if (type === 'delete') {
 				const existing = table.get(this.getPK(tableName, record));
+				const validationError = this.validate(tableName, 'delete', record);
 				this.log('delete looking for existing', { existing });
-				if (!existing) {
+
+				if (validationError) {
+					data = {
+						[selection]: null,
+					};
+					errors = [validationError];
+				} else if (!existing) {
 					data = {
 						[selection]: null,
 					};
@@ -991,7 +1128,8 @@ export function getDataStore({
 			'https://0.0.0.0/graphql';
 	}
 
-	const classes = initSchema(testSchema());
+	const schema = testSchema();
+	const classes = initSchema(schema);
 
 	const {
 		ModelWithBoolean,
@@ -1018,6 +1156,12 @@ export function getDataStore({
 		LegacyJSONComment,
 		CompositePKParent,
 		CompositePKChild,
+		BasicModel,
+		BasicModelWritableTS,
+		BasicModelRequiredTS,
+		ModelWithExplicitOwner,
+		ModelWithExplicitCustomOwner,
+		ModelWithMultipleCustomOwner,
 		ModelWithIndexes,
 	} = classes as {
 		ModelWithBoolean: PersistentModelConstructor<ModelWithBoolean>;
@@ -1044,11 +1188,18 @@ export function getDataStore({
 		LegacyJSONComment: PersistentModelConstructor<LegacyJSONComment>;
 		CompositePKParent: PersistentModelConstructor<CompositePKParent>;
 		CompositePKChild: PersistentModelConstructor<CompositePKChild>;
+		BasicModel: PersistentModelConstructor<BasicModel>;
+		BasicModelWritableTS: PersistentModelConstructor<BasicModelWritableTS>;
+		BasicModelRequiredTS: PersistentModelConstructor<BasicModelRequiredTS>;
+		ModelWithExplicitOwner: PersistentModelConstructor<ModelWithExplicitOwner>;
+		ModelWithExplicitCustomOwner: PersistentModelConstructor<ModelWithExplicitCustomOwner>;
+		ModelWithMultipleCustomOwner: PersistentModelConstructor<ModelWithMultipleCustomOwner>;
 		ModelWithIndexes: PersistentModelConstructor<ModelWithIndexes>;
 	};
 
 	return {
 		DataStore,
+		schema,
 		connectivityMonitor,
 		graphqlService,
 		simulateConnect,
@@ -1077,6 +1228,12 @@ export function getDataStore({
 		LegacyJSONComment,
 		CompositePKParent,
 		CompositePKChild,
+		BasicModel,
+		BasicModelWritableTS,
+		BasicModelRequiredTS,
+		ModelWithExplicitOwner,
+		ModelWithExplicitCustomOwner,
+		ModelWithMultipleCustomOwner,
 		ModelWithIndexes,
 	};
 }
@@ -1107,14 +1264,14 @@ export const DataStore: typeof DS = (() => {
 export declare class Model {
 	public readonly id: string;
 	public readonly field1: string;
-	public readonly optionalField1?: string;
+	public readonly optionalField1?: string | null;
 	public readonly dateCreated: string;
-	public readonly emails?: string[];
-	public readonly ips?: (string | null)[];
-	public readonly metadata?: Metadata;
-	public readonly logins?: Login[];
-	public readonly createdAt?: string;
-	public readonly updatedAt?: string;
+	public readonly emails?: string[] | null;
+	public readonly ips?: (string | null)[] | null;
+	public readonly metadata?: Metadata | null;
+	public readonly logins?: Login[] | null;
+	public readonly createdAt?: string | null;
+	public readonly updatedAt?: string | null;
 
 	constructor(init: ModelInit<Model>);
 
@@ -1296,6 +1453,42 @@ export declare class BasicModel {
 			draft: MutableModel<BasicModel>
 		) => MutableModel<BasicModel> | void
 	): BasicModel;
+}
+
+export declare class BasicModelWritableTS {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<BasicModelWritableTS, 'id'>;
+		readOnlyFields: never;
+	};
+	readonly id: string;
+	readonly body: string;
+	readonly createdAt?: string | null;
+	readonly updatedAt?: string | null;
+	constructor(init: ModelInit<BasicModelWritableTS>);
+	static copyOf(
+		source: BasicModelWritableTS,
+		mutator: (
+			draft: MutableModel<BasicModelWritableTS>
+		) => MutableModel<BasicModelWritableTS> | void
+	): BasicModelWritableTS;
+}
+
+export declare class BasicModelRequiredTS {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<BasicModelRequiredTS, 'id'>;
+		readOnlyFields: never;
+	};
+	readonly id: string;
+	readonly body: string;
+	readonly createdAt: string;
+	readonly updatedOn?: string | null;
+	constructor(init: ModelInit<BasicModelRequiredTS>);
+	static copyOf(
+		source: BasicModelRequiredTS,
+		mutator: (
+			draft: MutableModel<BasicModelRequiredTS>
+		) => MutableModel<BasicModelRequiredTS> | void
+	): BasicModelRequiredTS;
 }
 
 export declare class HasOneParent {
@@ -1643,6 +1836,64 @@ export declare class ModelWithBoolean {
 	): ModelWithBoolean;
 }
 
+export declare class ModelWithExplicitOwner {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<ModelWithExplicitOwner, 'id'>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly id: string;
+	readonly title: string;
+	readonly owner?: string | null;
+	readonly createdAt?: string | null;
+	readonly updatedAt?: string | null;
+	constructor(init: ModelInit<ModelWithExplicitOwner>);
+	static copyOf(
+		source: ModelWithExplicitOwner,
+		mutator: (
+			draft: MutableModel<ModelWithExplicitOwner>
+		) => MutableModel<ModelWithExplicitOwner> | void
+	): ModelWithExplicitOwner;
+}
+
+export declare class ModelWithExplicitCustomOwner {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<ModelWithExplicitCustomOwner, 'id'>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly id: string;
+	readonly title: string;
+	readonly customowner?: string | null;
+	readonly createdAt?: string | null;
+	readonly updatedAt?: string | null;
+	constructor(init: ModelInit<ModelWithExplicitCustomOwner>);
+	static copyOf(
+		source: ModelWithExplicitCustomOwner,
+		mutator: (
+			draft: MutableModel<ModelWithExplicitCustomOwner>
+		) => MutableModel<ModelWithExplicitCustomOwner> | void
+	): ModelWithExplicitCustomOwner;
+}
+
+export declare class ModelWithMultipleCustomOwner {
+	readonly [__modelMeta__]: {
+		identifier: OptionallyManagedIdentifier<ModelWithMultipleCustomOwner, 'id'>;
+		readOnlyFields: 'createdAt' | 'updatedAt';
+	};
+	readonly id: string;
+	readonly title: string;
+	readonly customownerOne?: string | null;
+	readonly customownerTwo?: string | null;
+	readonly createdAt?: string | null;
+	readonly updatedAt?: string | null;
+	constructor(init: ModelInit<ModelWithMultipleCustomOwner>);
+	static copyOf(
+		source: ModelWithMultipleCustomOwner,
+		mutator: (
+			draft: MutableModel<ModelWithMultipleCustomOwner>
+		) => MutableModel<ModelWithMultipleCustomOwner> | void
+	): ModelWithMultipleCustomOwner;
+}
+
 export declare class ModelWithIndexes {
 	public readonly id: string;
 	public readonly stringField?: string;
@@ -1971,7 +2222,7 @@ export function testSchema(): Schema {
 						name: 'profileID',
 						isArray: false,
 						type: 'ID',
-						isRequired: true,
+						isRequired: false,
 						attributes: [],
 					},
 					profile: {
@@ -2305,6 +2556,119 @@ export function testSchema(): Schema {
 						type: 'key',
 						properties: {
 							fields: ['id'],
+						},
+					},
+				],
+			},
+			BasicModelWritableTS: {
+				name: 'BasicModelWritableTS',
+				fields: {
+					id: {
+						name: 'id',
+						isArray: false,
+						type: 'ID',
+						isRequired: true,
+						attributes: [],
+					},
+					body: {
+						name: 'body',
+						isArray: false,
+						type: 'String',
+						isRequired: true,
+						attributes: [],
+					},
+					createdAt: {
+						name: 'createdAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: false,
+					},
+					updatedAt: {
+						name: 'updatedAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: false,
+					},
+				},
+				syncable: true,
+				pluralName: 'BasicModelWritableTimestampss',
+				attributes: [
+					{
+						type: 'model',
+						properties: {},
+					},
+					{
+						type: 'key',
+						properties: {
+							fields: ['id'],
+						},
+					},
+				],
+			},
+			/**
+			 * timestamps: {
+			 * 	updatedAt: "updatedOn"
+			 * }
+			 */
+			BasicModelRequiredTS: {
+				name: 'BasicModelRequiredTS',
+				fields: {
+					id: {
+						name: 'id',
+						isArray: false,
+						type: 'ID',
+						isRequired: true,
+						attributes: [],
+					},
+					body: {
+						name: 'body',
+						isArray: false,
+						type: 'String',
+						isRequired: true,
+						attributes: [],
+					},
+					createdAt: {
+						name: 'createdAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: true, // `AWSDateTime!` (intentionally required)
+						attributes: [],
+					},
+					updatedOn: {
+						name: 'updatedOn', //
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false, // `AWSDateTime` (intentionally optional)
+						attributes: [],
+					},
+				},
+				syncable: true,
+				pluralName: 'BasicModelRequiredTS',
+				attributes: [
+					{
+						type: 'model',
+						properties: {
+							timestamps: {
+								updatedAt: 'updatedOn',
+							},
+						},
+					},
+					{
+						type: 'auth',
+						properties: {
+							rules: [
+								{
+									provider: 'userPools',
+									ownerField: 'owner',
+									allow: 'owner',
+									identityClaim: 'cognito:username',
+									operations: ['create', 'update', 'delete', 'read'],
+								},
+							],
 						},
 					},
 				],
@@ -3485,6 +3849,212 @@ export function testSchema(): Schema {
 					{
 						type: 'model',
 						properties: {},
+					},
+				],
+			},
+			ModelWithExplicitOwner: {
+				name: 'ModelWithExplicitOwner',
+				fields: {
+					id: {
+						name: 'id',
+						isArray: false,
+						type: 'ID',
+						isRequired: true,
+						attributes: [],
+					},
+					title: {
+						name: 'title',
+						isArray: false,
+						type: 'String',
+						isRequired: true,
+						attributes: [],
+					},
+					owner: {
+						name: 'owner',
+						isArray: false,
+						type: 'String',
+						isRequired: false,
+						attributes: [],
+					},
+					createdAt: {
+						name: 'createdAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: true,
+					},
+					updatedAt: {
+						name: 'updatedAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: true,
+					},
+				},
+				syncable: true,
+				pluralName: 'ModelWithExplicitOwners',
+				attributes: [
+					{
+						type: 'model',
+						properties: {},
+					},
+					{
+						type: 'auth',
+						properties: {
+							rules: [
+								{
+									provider: 'userPools',
+									ownerField: 'owner',
+									allow: 'owner',
+									identityClaim: 'cognito:username',
+									operations: ['create', 'update', 'delete', 'read'],
+								},
+							],
+						},
+					},
+				],
+			},
+			ModelWithExplicitCustomOwner: {
+				name: 'ModelWithExplicitCustomOwner',
+				fields: {
+					id: {
+						name: 'id',
+						isArray: false,
+						type: 'ID',
+						isRequired: true,
+						attributes: [],
+					},
+					title: {
+						name: 'title',
+						isArray: false,
+						type: 'String',
+						isRequired: true,
+						attributes: [],
+					},
+					customowner: {
+						name: 'customowner',
+						isArray: false,
+						type: 'String',
+						isRequired: false,
+						attributes: [],
+					},
+					createdAt: {
+						name: 'createdAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: true,
+					},
+					updatedAt: {
+						name: 'updatedAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: true,
+					},
+				},
+				syncable: true,
+				pluralName: 'ModelWithExplicitCustomOwners',
+				attributes: [
+					{
+						type: 'model',
+						properties: {},
+					},
+					{
+						type: 'auth',
+						properties: {
+							rules: [
+								{
+									provider: 'userPools',
+									ownerField: 'customowner',
+									allow: 'owner',
+									identityClaim: 'cognito:username',
+									operations: ['create', 'update', 'delete', 'read'],
+								},
+							],
+						},
+					},
+				],
+			},
+			ModelWithMultipleCustomOwner: {
+				name: 'ModelWithMultipleCustomOwner',
+				fields: {
+					id: {
+						name: 'id',
+						isArray: false,
+						type: 'ID',
+						isRequired: true,
+						attributes: [],
+					},
+					title: {
+						name: 'title',
+						isArray: false,
+						type: 'String',
+						isRequired: true,
+						attributes: [],
+					},
+					customownerOne: {
+						name: 'customownerOne',
+						isArray: false,
+						type: 'String',
+						isRequired: false,
+						attributes: [],
+					},
+					customownerTwo: {
+						name: 'customownerTwo',
+						isArray: false,
+						type: 'String',
+						isRequired: false,
+						attributes: [],
+					},
+					createdAt: {
+						name: 'createdAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: true,
+					},
+					updatedAt: {
+						name: 'updatedAt',
+						isArray: false,
+						type: 'AWSDateTime',
+						isRequired: false,
+						attributes: [],
+						isReadOnly: true,
+					},
+				},
+				syncable: true,
+				pluralName: 'ModelWithMultipleCustomOwners',
+				attributes: [
+					{
+						type: 'model',
+						properties: {},
+					},
+					{
+						type: 'auth',
+						properties: {
+							rules: [
+								{
+									provider: 'userPools',
+									ownerField: 'customownerOne',
+									allow: 'owner',
+									identityClaim: 'cognito:username',
+									operations: ['create', 'update', 'delete', 'read'],
+								},
+								{
+									provider: 'userPools',
+									ownerField: 'customownerTwo',
+									allow: 'owner',
+									identityClaim: 'cognito:username',
+									operations: ['create', 'read'],
+								},
+							],
+						},
 					},
 				],
 			},