+++ packages/api-rest/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.18](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@3.0.17...@aws-amplify/api-rest@3.0.18) (2023-03-08)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+
+
+
+
+## [3.0.17](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@3.0.16...@aws-amplify/api-rest@3.0.17) (2023-03-06)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+
+
+
+
 ## [3.0.16](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@3.0.15...@aws-amplify/api-rest@3.0.16) (2023-02-24)
 
 **Note:** Version bump only for package @aws-amplify/api-rest